WIP: multithreded test; debuging no sub slots error
