+++ drivers/nvme-ssd/tests.h
@@ -1,3 +1,6 @@
+#ifndef __TESTS_H__
+#define __TESTS_H__
+
 #include <common/dump_utils.h>
 
 class TestBlockWriter
@@ -68,3 +71,4 @@ public:
   }
 
 };
+#endif
